core code updates
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base@4966 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ trunk/src/org/hsqldb/index/Index.java
@@ -42,7 +42,7 @@ import org.hsqldb.types.Type;
 /**
  *
  * @author Fred Toussi (fredt@users dot sourceforge.net)
- * @version 2.0.1
+ * @version 2.2.9
  * @since 1.9.0
  */
 public interface Index extends SchemaObject {
@@ -133,9 +133,9 @@ public interface Index extends SchemaObject {
     /**
      * Returns the node count.
      */
-    public int size(Session session, PersistentStore store);
+    public long size(Session session, PersistentStore store);
 
-    public int sizeUnique(PersistentStore store);
+    public long sizeUnique(PersistentStore store);
 
     public boolean isEmpty(PersistentStore store);
